feat: add section numbering to PDF bookmark outline
As of Typst 0.13.1, there is no section numberin in PDF bookmarks.
The way to work around this is to hide the actual heading from the
bookmarks, and create an invisible, bookmarked heading with prepended
numberings just above it.

These invisible headings are placed before the visible, actual headings
BUT after the pre-heading page/line breaks for accurate PDF link redirection.

All these are refactored into two files, hacks.typ and headings.typ
for easy management.